# [`Optimize`]: `Murmur128` is a noncryptographic hash aglorithm