# Description

Same as the  murmur32, murmur128 is a noncryptographic hash aglorithm.
So it shouldn't be implemented as a cryptographic hash (This would be misleading.):

This PR:
1. Implements murmur128 as a noncryptographic hash aglorithm;
2. Optimizes murmur128 performance in some cases;

A simple benchmark:

Before:
```
.NET SDK 9.0.203
  [Host]     : .NET 9.0.4 (9.0.425.16305), Arm64 RyuJIT AdvSIMD
  DefaultJob : .NET 9.0.4 (9.0.425.16305), Arm64 RyuJIT AdvSIMD


| Method    | Mean     | Error    | StdDev   |
|---------- |---------:|---------:|---------:|
| Murmur128 | 81.65 us | 1.461 us | 1.295 us |
```

After:
```
.NET SDK 9.0.203
  [Host]     : .NET 9.0.4 (9.0.425.16305), Arm64 RyuJIT AdvSIMD
  DefaultJob : .NET 9.0.4 (9.0.425.16305), Arm64 RyuJIT AdvSIMD


| Method                | Mean     | Error    | StdDev   |
|---------------------- |---------:|---------:|---------:|
| Murmur128_ComputeHash | 40.74 us | 0.097 us | 0.090 us |
```

Fixes # (issue)

## Type of change

<!-- Please delete options that are not relevant. -->

- [x] Optimization (the change is only an optimization)
- [ ] Style (the change is only a code style for better maintenance or standard purpose)
- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [ ] This change requires a documentation update

# Checklist:

- [x] My code follows the style guidelines of this project
- [x] I have performed a self-review of my code
- [x] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my feature works
- [x] New and existing unit tests pass locally with my changes
- [ ] Any dependent changes have been merged and published in downstream modules